+++ brewtils/schema_parser.py
@@ -23,6 +23,7 @@ class SchemaParser(object):
     _models = {
         "ChoicesSchema": brewtils.models.Choices,
         "CommandSchema": brewtils.models.Command,
+        "ConnectionSchema": brewtils.models.Connection,
         "CronTriggerSchema": brewtils.models.CronTrigger,
         "DateTriggerSchema": brewtils.models.DateTrigger,
         "EventSchema": brewtils.models.Event,
@@ -100,6 +101,22 @@ class SchemaParser(object):
         """
         return cls.parse(
             command, brewtils.models.Command, from_string=from_string, **kwargs
+        )
+
+    @classmethod
+    def parse_connection(cls, connection, from_string=False, **kwargs):
+        """Convert raw JSON string or dictionary to a command model object
+
+        Args:
+            connection: The raw input
+            from_string: True if input is a JSON string, False if a dictionary
+            **kwargs: Additional parameters to be passed to the Schema (e.g. many=True)
+
+        Returns:
+            A Command object
+        """
+        return cls.parse(
+            connection, brewtils.models.Connection, from_string=from_string, **kwargs
         )
 
     @classmethod
@@ -493,6 +510,26 @@ class SchemaParser(object):
             command,
             to_string=to_string,
             schema_name=brewtils.models.Command.schema,
+            **kwargs
+        )
+
+    @classmethod
+    def serialize_connection(cls, connection, to_string=True, **kwargs):
+        """Convert a connection model into serialized form
+
+        Args:
+            connection: The connection object(s) to be serialized
+            to_string: True to generate a JSON-formatted string, False to generate a
+                dictionary
+            **kwargs: Additional parameters to be passed to the Schema (e.g. many=True)
+
+        Returns:
+            Serialized representation of connection
+        """
+        return cls.serialize(
+            connection,
+            to_string=to_string,
+            schema_name=brewtils.models.Connection.schema,
             **kwargs
         )