+++ brewtils/schemas.py
@@ -35,6 +35,7 @@ __all__ = [
     "IntervalTriggerSchema",
     "CronTriggerSchema",
     "FileTriggerSchema",
+    "ConnectionSchema",
     "GardenSchema",
     "OperationSchema",
     "UserSchema",
@@ -348,6 +349,8 @@ class RequestSchema(RequestTemplateSchema):
     )
     has_parent = fields.Bool(allow_none=True)
     requester = fields.String(allow_none=True)
+    source_garden = fields.String(allow_none=True)
+    target_garden = fields.String(allow_none=True)
 
 
 class StatusInfoSchema(BaseSchema):
@@ -486,15 +489,33 @@ class FileTriggerSchema(BaseSchema):
     callbacks = fields.Dict(fields.Bool(), allow_none=True)
 
 
+class ConnectionSchema(BaseSchema):
+    api = fields.Str(allow_none=True)
+    status = fields.Str(allow_none=True)
+    status_info = fields.Nested("StatusInfoSchema", allow_none=True)
+    config = fields.Dict(allow_none=True)
+
+
 class GardenSchema(BaseSchema):
     id = fields.Str(allow_none=True)
     name = fields.Str(allow_none=True)
     status = fields.Str(allow_none=True)
     status_info = fields.Nested("StatusInfoSchema", allow_none=True)
     connection_type = fields.Str(allow_none=True)
-    connection_params = fields.Dict(allow_none=True)
+    receiving_connections = fields.Nested(
+        "ConnectionSchema", many=True, allow_none=True
+    )
+    publishing_connections = fields.Nested(
+        "ConnectionSchema", many=True, allow_none=True
+    )
     namespaces = fields.List(fields.Str(), allow_none=True)
     systems = fields.Nested("SystemSchema", many=True, allow_none=True)
+    has_parent = fields.Bool(allow_none=True)
+    parent = fields.Str(allow_none=True)
+    children = fields.Nested(
+        "self", exclude=("parent"), many=True, default=None, allow_none=True
+    )
+    metadata = fields.Dict(allow_none=True)
 
 
 class GardenDomainIdentifierSchema(BaseSchema):
@@ -560,6 +581,8 @@ class OperationSchema(BaseSchema):
 
     target_garden_name = fields.Str(allow_none=True)
     source_garden_name = fields.Str(allow_none=True)
+    source_api = fields.Str(allow_none=True)
+
     operation_type = fields.Str(allow_none=True)
 
 
@@ -621,6 +644,7 @@ model_schema_map.update(
     {
         "Choices": ChoicesSchema,
         "Command": CommandSchema,
+        "Connection": ConnectionSchema,
         "CronTrigger": CronTriggerSchema,
         "DateTrigger": DateTriggerSchema,
         "Event": EventSchema,